Moving vcharts inside candela as util/vega
