+++ src/candela/components/Box/index.js
@@ -1,4 +1,4 @@
-import vcharts from '../../../vcharts';
+import vega from '../../util/vega';
 import spec from './spec.json';
 
 export default class Box {
@@ -35,7 +35,7 @@ export default class Box {
   }
 
   constructor (el, options) {
-    this.chart = vcharts.chart(spec, el, options);
+    this.chart = vega.chart(spec, el, options);
   }
 
   render () {
