+++ src/candela/components/LineChart/index.js
@@ -1,5 +1,5 @@
 import VisComponent from '../../VisComponent';
-import vcharts from '../../../vcharts';
+import vega from '../../util/vega';
 import spec from './spec.json';
 
 export default class LineChart extends VisComponent {
@@ -26,7 +26,7 @@ export default class LineChart extends VisComponent {
   }
 
   data (data) {
-    this.chart = vcharts.chart(spec, this.el, {
+    this.chart = vega.chart(spec, this.el, {
       series: [
         {
           name: 'series1',
