+++ src/candela/components/ScatterMatrix/index.js
@@ -1,4 +1,4 @@
-import vcharts from '../../../vcharts';
+import vega from '../../util/vega';
 import spec from './spec.json';
 
 export default class ScatterMatrix {
@@ -35,7 +35,7 @@ export default class ScatterMatrix {
   }
 
   constructor (el, options) {
-    this.chart = vcharts.chart(spec, el, options);
+    this.chart = vega.chart(spec, el, options);
   }
 
   render () {
