+++ src/candela/util/vega/index.js
