+++ src/candela/util/vega/test/chart.js
